% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbiologging-package.R
\docType{package}
\name{openbiologging-package}
\alias{openbiologging}
\alias{openbiologging-package}
\title{openbiologging: Open Data Practices in Biologging Research}
\description{
Provides data from a systematic literature review assessing open data
practices in biologging research. See \code{\link{biologging}} and
\code{\link{taxa}} for the datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/haylee360/openbiologging}
  \item Report bugs at \url{https://github.com/haylee360/openbiologging/issues}
}

}
\author{
\strong{Maintainer}: Max Czapanskiy \email{maxczap@ucsb.edu}

Authors:
\itemize{
  \item Max Czapanskiy \email{maxczap@ucsb.edu}
  \item Haylee Oyler \email{hayleeoyler360@gmail.com}
}

}
\keyword{internal}
